requirements.txt
Flask
gunicorn
Pillow
reportlab